Refactor display tasks feature
